+++ androidApp/core/src/main/java/io/github/nadzwyczajnaGrupaRobocza/texaspoker/screen/GameScreen.kt
@@ -5,24 +5,52 @@ import com.badlogic.gdx.assets.AssetManager
 import com.badlogic.gdx.graphics.OrthographicCamera
 import com.badlogic.gdx.graphics.g2d.Batch
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer
+import com.badlogic.gdx.utils.Array
+import io.github.nadzwyczajnaGrupaRobocza.texaspoker.actors.CommunityCardsActor
+import io.github.nadzwyczajnaGrupaRobocza.texaspoker.actors.GameActor
+import io.github.nadzwyczajnaGrupaRobocza.texaspoker.actors.PlayersRingActor
+import io.github.nadzwyczajnaGrupaRobocza.texaspoker.actors.TableActor
 import io.github.nadzwyczajnaGrupaRobocza.texaspoker.ecs.systems.ShapeRenderingSystem
 import io.github.nadzwyczajnaGrupaRobocza.texaspoker.ecs.systems.SpriteRenderingSystem
-import io.github.nadzwyczajnaGrupaRobocza.texaspoker.ecs.systems.TableAnimationSystem
 import ktx.app.KtxScreen
+import kotlin.collections.arrayListOf
 
 class GameScreen(
     private val batch: Batch,
     private val assets: AssetManager,
     private val camera: OrthographicCamera,
     private val engine: PooledEngine,
-    private val shape_renderer : ShapeRenderer
+    private val shape_renderer: ShapeRenderer
 ) : KtxScreen {
 
+    // All classes that depends on textures cannot be created at construction time
+    // Textures are load just before the show() function is called
+    private var actors = arrayListOf<GameActor?>()
+
     override fun render(delta: Float) {
+        updateGameScene(delta)
         engine.update(delta)
     }
 
     override fun show() {
+        createGameScene()
+
+        setupEntityComponentSystems()
+    }
+
+    private fun updateGameScene(delta: Float) {
+        for (actor in actors) {
+            actor?.update(delta)
+        }
+    }
+
+    private fun createGameScene() {
+        actors.add(TableActor(assets, engine, camera.viewportWidth, camera.viewportHeight))
+        actors.add(CommunityCardsActor(assets, engine, camera.viewportWidth, camera.viewportHeight))
+        actors.add(PlayersRingActor(assets, engine, camera.viewportWidth, camera.viewportHeight))
+    }
+
+    private fun setupEntityComponentSystems() {
         engine.apply {
             addSystem(
                 SpriteRenderingSystem(
@@ -30,7 +58,6 @@ class GameScreen(
                     camera
                 )
             )
-            addSystem(TableAnimationSystem(assets, camera.viewportWidth, camera.viewportHeight))
             addSystem(ShapeRenderingSystem(shape_renderer))
         }
     }
